requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-WTF
Werkzeug
Flask-Login
python-dotenv